refactor: Extract message saving to helper function
The code for saving messages to the database has been extracted into a helper function, `saveMessage`, and replaced two instances of similar code in `POST`. This is done for code readability and to keep the DRY (Don't Repeat Yourself) principle in mind. Moreover, an error check has been added for an empty prompt which will return a 500 status code in such case.