feat: Add title to chat and initial messages in chat page
This commit adds a 'title' field in the chat model in prisma schema, with the corresponding change in the creation of new chats in chat route. This is to display a short prompt as the title of each chat when they are listed.

The initial messages of a chat are also now fetched in the chat page, with corresponding changes in the chat component and providers. This is to display the previous chat messages when a user revisits a chat.

A new utility function 'shorten' has been added for truncating long strings, ideal for creating chat titles from prompts. A sidebar component was also added to the header.

Please, note that query caching libraries were introduced and the 'Chat' component now accepts 'initialMessages' to facilitate these changes.