feat: Implement selective sign-in based on email whitelist
This commit adds a new feature to the authentication method in `auth.ts` to check a user's email against a list of allowed emails for additional security. A user with an email that's not on the allowed list will receive an error message. On the sign-in page, a warning message has been added prompting non-invited users not to attempt signing in.