refactor: Update JWT import and extend session types
This refactor changes the source of the JWT import in lib/auth.ts. It is required due to the change in where the JWT is defined, improving organization and code flow. The JWT is now imported from 'lib/session' instead of 'next-auth/jwt'. Furthermore, Session and JWT interfaces are extended in lib/session.ts to include the 'id' property, providing more specificity in types and aiding in error prevention during development. The inclusion of potential null value for the session also defends against possible runtime errors.