misc: Upgrade project dependencies to latest versions
The change updates the version number of multiple libraries used in the project inside the package.json file. This was necessary to ensure that we are using the latest and most secure versions of the libraries, benefit from the bug fixes, performance improvements and new features that these new versions bring, and keep the project up-to-date.