feature: Add System Messages and User Profile
This commit adds multiple functionalities to improve the user interface and provide more user-specific features.

A new feature for System Messages has been added. Users can now create and manage System Messages, this change required a new table "SystemMessage" to be added to the database schema.

ProfilePage was introduced, where a user can view and change their System Message. For this, a "profile" route is implemented, along with a custom hook for fetching and updating the user's System Message from the server.

Also, the User's Usage information was moved to the UserMenu from the ProfilePage. UsageDisplay component is created for the same.

Several other minor changes were introduced including updates to routes, hooks and utility functions to accommodate these new features.