misc: Upgrade packages and Yarn version in project
This commit updates the versions of multiple dependencies in package.json to ensure compatibility with recent updates and maintain project continuity. It includes both major and minor versions' upgrade. Yarn has also been updated from version 3.6.3 to 4.0.0 as specified in .yarnrc.yml to benefit from the latest features and improvements.