refactor: Upgrade @tanstack/query-core and react-query versions
Updated versions of @tanstack/query-core and react-query from 4.36.1 to 5.0.5 in package.json and yarn.lock files. This brings the project up-to-date with the latest improvements and fixes in these libraries. Also, various queries were refactored to align with the updated react-query API, which includes usage of 'queryKey' and 'queryFn'. 'isLoading' prop was replaced by 'isPending' to match the new API.