refactor: Renamed AllowedUsers to AllowedUsersDisplay and added key prop for each item
The file allowed-users.tsx has been moved to the allowed-users directory and renamed to allowed-users-display.tsx, thus better outlining the responsibility of this component. The component was previously called 'AllowedUsers', it is now called 'AllowedUsersDisplay' to increase overall clarity and understanding about the component's purpose.

In addition, a react key prop has been added to each child component created in the allowedUsers.map function call in the component. This improves list rendering and helps React with diffing the child components during re-renders. References to AllowedUsers have been updated in the app/profile/page.tsx to reflect this change.