refactor: Remove unused code and improve user interface interaction
Removed the unused `useForm` hook in `delete-allowed-user-form.tsx`, as it was unnecessary for the functioning of the delete operation. Also, the UI for `system-message-form.tsx` and `add-allowed-user-form.tsx` were improved to give feedback on submission state, hence improving user interaction. The width of the items within `allowed-users-display.tsx` has been slightly increased to allow for better readability and display of user information. This makes for a more user-friendly experience overall.