feat: Fetch user usage data on initial render
Updated the UsageDisplay, Header and UserMenu components to fetch user's usage data on initial render for a better user experience. The UserMenu component now receives user usage data as a prop. This removes the need for the client to wait for a refetch to view their usage data. Changes in links and revalidation paths from '/settings' to '/profile' have also been made accordingly for same user experience reasons.