misc: Change Roboto_Mono font to JetBrains_Mono in ui/fonts
The font used for the mono style in our UI has been updated from Roboto_Mono to JetBrains_Mono. This change aligns with the new branding guidelines for an updated look and feel of the application. JetBrains_Mono is known for its readability and clean lines, providing a better user experience.